Refactoring on prop files % config files
