--- account-service/src/main/java/pl/piomin/microservices/account/AccountConfiguration.java
@@ -1,17 +1,0 @@
-package pl.piomin.microservices.account;
-
-import org.springframework.cloud.sleuth.sampler.AlwaysSampler;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan("pl.piomin.microservices.account")
-public class AccountConfiguration {
-
-	@Bean
-	public AlwaysSampler defaultSampler() {
-	  return new AlwaysSampler();
-	}
-	
-}
