+++ account-service/src/main/java/pl/piomin/microservices/account/config/AccountConfiguration.java
@@ -1,0 +1,17 @@
+package pl.piomin.microservices.account.config;
+
+import org.springframework.cloud.sleuth.sampler.AlwaysSampler;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan("pl.piomin.microservices.account")
+public class AccountConfiguration {
+
+	@Bean
+	public AlwaysSampler defaultSampler() {
+	  return new AlwaysSampler();
+	}
+	
+}
