+++ customer-service/src/main/resources/application.yml
@@ -1,14 +1,14 @@
-server:  
+server:
   port: ${PORT:3333}
 
-eureka:  
+eureka:
   client:
     serviceUrl:
-      defaultZone: ${DISCOVERY_URL:http://localhost:8761}/eureka/
+      defaultZone: http://localhost:8761}/eureka/
   instance:
     leaseRenewalIntervalInSeconds: 1
     leaseExpirationDurationInSeconds: 2
-          
+
 ribbon:
   eureka:
     enabled: true
