+++ customer-service/src/main/resources/bootstrap.yml
@@ -1,4 +1,4 @@
-spring:  
+spring:
   application:
     name: customer-service
   logging:
@@ -6,4 +6,4 @@ spring:
       console: "%clr(%d{yyyy-MM-dd HH:mm:ss}){faint} %clr(${LOG_LEVEL_PATTERN:-%5p}) %clr([${springAppName:-},%X{X-B3-TraceId:-},%X{X-B3-SpanId:-},%X{X-Span-Export:-}]){yellow} %clr(${PID:- }){magenta} %clr(---){faint} %clr([%15.15t]){faint} %clr(%-40.40logger{39}){cyan} %clr(:){faint} %m%n${LOG_EXCEPTION_CONVERSION_WORD:-%wEx}"
     level:
       org.springframework: WARN
-      org.hibernate: WARN    
+      org.hibernate: WARN
