+++ gateway-service/src/main/java/pl/piomin/microservices/edge/config/CustomerConfiguration.java
@@ -1,0 +1,17 @@
+package pl.piomin.microservices.edge.config;
+
+import org.springframework.cloud.sleuth.sampler.AlwaysSampler;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan("pl.piomin.microservices.edge")
+public class CustomerConfiguration {
+
+    @Bean
+    public AlwaysSampler defaultSampler() {
+        return new AlwaysSampler();
+    }
+
+}
