+++ zipkin-service/src/main/resources/application.yml
@@ -1,2 +1,2 @@
-server:  
+server:
   port: ${PORT:9411}
